+++ CHANGELOG.md
@@ -4,10 +4,13 @@
   - `RegExp.prototype[@@match]`
   - `RegExp.prototype[@@replace]`
 - Improved handling of `RegExp` flags in polyfills of some methods in engines without proper support of `RegExp.prototype.flags` and without polyfill of this getter
+- Added feature detection for [a bug](https://bugs.webkit.org/show_bug.cgi?id=288595) that occurs when `this` is updated while `Set.prototype.difference` is being executed
 - Compat data improvements:
   - [`Error.isError`](https://github.com/tc39/proposal-is-error) marked not supported in Node because of [a bug](https://github.com/nodejs/node/issues/56497)
   - Added [Deno 2.3](https://github.com/denoland/deno/releases/tag/v2.3.0) compat data mapping
   - Updated Electron 37 compat data mapping
+  - `Set.prototype.difference` marked not supported in Safari because of [a bug](https://bugs.webkit.org/show_bug.cgi?id=288595)
+  - `Set.prototype.difference` marked as supported from Bun 1.2.5 because of [a bug](https://bugs.webkit.org/show_bug.cgi?id=288595) was fixed
 
 ##### [3.42.0 - 2025.04.30](https://github.com/zloirock/core-js/releases/tag/v3.42.0)
 - Changes [v3.41.0...v3.42.0](https://github.com/zloirock/core-js/compare/v3.41.0...v3.42.0) (142 commits)